Use device IDs as passed around value
By using the ID we can actually get the picker to update as expected and
threading that through as a Binding ensures that the change bubbles all
the way up. Unfortunately (though this could be changed) the
AudioDeviceManager uses AudioDevice as the interface and not the ID
alone. So we have to do a little transformation when the change happens
in a manual `set` closure. Not ideal but fixable probably.